add header.h and function comments
